+++ CHANGELOG.md
@@ -15,6 +15,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## [0.1.x]
 
+- Added support for setting file and editor attributes, to facilitate use of external editor via setting inspector vars or via setting `editor <path>` and `edit <filename>`. Also double-clicking on the `chuck~` object will open the current file in the specified external editor.
+
+- Fixed `chuck~` object args to allow just the number of channels as args.
+
 - Added midi player to .maxhelp to demonstrate midi file playback of `chuck~` from Max.
 
 - Added `loglevel` / `loglevel <int>` message for setting chuck log level 0-10. Dropped redundant previous `debug` switch/attribute.